requirements.txt
aioboto3==15.1.0
aiobotocore==2.24.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==2.4.1
attrs==25.3.0
azure-cognitiveservices-speech==1.45.0
azure-core==1.35.0
boto3==1.39.11
botocore==1.39.11
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.116.1
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.30.0
ipython_pygments_lexers==1.0.0
jedi==0.19.2
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain-core==0.3.74
langchain-openai==0.3.30
langsmith==0.4.14
matplotlib-inline==0.1.7
multidict==6.6.4
nest_asyncio==1.6.0
numpy==1.26.4
openai==1.99.9
orjson==3.11.2
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.3.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.18.0
python-dateutil==2.9.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
pyzmq==26.2.0
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
s3transfer==0.13.1
setuptools==78.1.1
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.2
stack_data==0.6.3
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.11.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==10.4
wheel==0.45.1
wrapt==1.17.3
yarl==1.20.1
zipp==3.21.0
zstandard==0.23.0